src/bundle: fix assert handling in uint64 read/write functions
The output_stream_write_uint64_all() and input_stream_read_uint64_all()
methods had GError handling in theory, but actually, in most cases this
will be pointless.
The reason is that there is a 'g_assert()' that will fail before being
able to indicate an error as the function return code and lead to hard
aborts like:

| Bail out! rauc:ERROR:src/bundle.c:479:input_stream_read_uint64_all: assertion failed: (bytes_read == sizeof(tmp))

Fix this by directly returning FALSE in case of an error in the
read/write calls and use the g_assert() only if no error has been
detected before.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>