src/slot: do not clean data directory for slot status 'pending'
In 67ccd898 ("src/install: ensure invalidating slot status before
updating"), a new 'pending' state for slots was added to allow
identifying slots that were left in an undefined state (mainly by rauc
crashing during installation).

In 'pending' state, the slot does not have a checksum.

Since we currently use the slot status saving for calling the directory
cleanup, the cleanup will also be called when saving the 'pending'
state (with empty slot checksum information).

As a result, the cleanup will also remove current and still valid
checksum-based hash directories used for adaptive mode
'block-hash-index' so that these need to be regenerated on each
installation. Depending on the platform and storage, this can be quite
time-consuming.

For now, fix this by explicitly leaving the data directory untouched in
case of a 'pending' slot.

A later option is to re-evaluate the slot status handling in general and
then see if the data directory cleanup should be coupled or called
somewhere else.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>