contrib: rauc-client: simple update service for rauc
This service implements a simple remote update service for rauc.
The service will check at time or intervals if there is a new
release available on the select update channel url.
If the version is newer of the current installed version,
it calls `rauc install` to install the update and reboot.

The service is simple enough to be compatible with busybox posix
shell and can be easily integrated in any environment.
It depends only on curl and posix shell.

Signed-off-by: Angelo Compagnucci <angelo.compagnucci@gmail.com>