test: pass meson build directory to tests
So far we have used a hack to guess the meson build directory in the
tests. This works fine as long as the build directory is just "/build"
and located in the subdirectory of tests.

However, this is not always the case, for example when running tests
as part of "meson dist".

Fix this known limitation by passing the build directory as
MESON_BUILD_DIR env variable to the pytest test scripts.

We use the rauc_build_root variable from rauc's base meson.build for
this to get the base build path (instead of the test/ build path).

From conftest.py use MESON_BUILD_DIR to set meson_build now and use
os.path.join() to construct the paths correctly.

Use meson_build also for setting PATH.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>