src: add missing "PRIORITY" field to structured logging arrays
In contrast to g_log_structured(), g_log_structured_array() takes a
'log_level' as input argument but uses this for filtering only.

The "PRIORITY" field must already be present in the array fields
provided.

Note that the messages are shown in systemd with default priority
anyway. A difference can only be observed when using "journalctl" with
"-o verbose".

Fixes: #1470

Reported-by: Andreas Ufert
Signed-off-by: Enrico Joerns <ejo@pengutronix.de>