src/context: convert 'system_variant' from char to array
Convert the system_variant stored in the context from a simple string,
to an array of strings. This is done to make use of the full
list/hiearchy that can be provided in a devicetree compatible.

To quote the kernel docs [1]:
The ‘compatible’ property contains a sorted list of strings starting
with the exact name of the machine, followed by an optional list of
boards it is compatible with sorted from most compatible to least.

Link [1]: https://www.kernel.org/doc/html/latest/devicetree/usage-model.html#platform-identification
Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>