src/update_handler: fix occasional broken pipe error on install
RAUC is splicing data from the input file (a .tar inside the bundle) to
stdin of a `tar` process it spawns to untar it.  This is done in chunks
in a while() loop, where the loop is exited if splice() ever returns 0
(presumably because the input FD hit EOF).

Under high CPU load, I've seen the last chunk of data reach the `tar`
process and for it to complete the untarring and exit successfully
before the loop in the installer process goes around one last time to
call splice() and get the 0 return value.  If this happens, when it
finally does call splice() one more time, the destination (`tar`) is
gone, so splice() reports a broken pipe and the process aborts, even
though `tar` completed successfully.

The fix is to add a check to the while() loop to break if all of the
data has been spliced, even if the last splice() call returned a
non-zero value.

Signed-off-by: David Hollister <david.hollister@ptxtrimble.com>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>