src/main: avoid tabs in status output
Using \t in the strings makes them hard to align in the source code and
results in tab characters in the output, which tends to disrupt the
alignment when copying the output somewhere else. So, keep it simple by
aligning using spaces in the source as well.

This also fixes inconsistent alignment when using --detailed.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>