install: expose manifest meta-data in install-check hook environment
This uses the same variable names (and implementation) as 'rauc info'.
Only the install-check hook is updated for now, as the slot hooks
currently don't have access to the manifest. We'd need to pass the
manifest or prepared variables down through all the update handles, but
that's a larger change.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>