qemu-test-init: create coverage data helper scripts
When working on the tests, it's often useful to update the coverage
report without exiting and restarting the qemu environment. To support
that, the collected coverage needs to written back to the host. Make the
helper function a standalone script to support that use-case.

Also add a clear_gcov_data script to use after rebuilding the binaries.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>