add prevent-fallback system option
When enabled, marking the booted slot as good will result in also
marking the others as bad. This can be used to ensure not booting
into the previous used slot after installing a new rootfs.

The option be set in the [system] section like this:

```
[system]
compatible=My BSP
bootloader=barebox
bundle-formats=-plain
data-directory=/data/rauc
prevent-fallback=true
....
```

In order to mark all other slots as bad we call the mark_run
recursively. The messages of both calls (mark-good and mark-bad) will be
concatenated, whereas for the slot names we only preserve the name of the
slot that was marked good.

Signed-off-by: Fabian Pfitzner <f.pfitzner@pengutronix.de>