support locating system.conf in /usr and /run
Traditionally, programs look up their configuration in /etc and this is
also what RAUC has been doing so far.

The modern pattern is to consult configuration files at a number of
locations. The default distro-provided config goes into /usr/lib,
which may be even read-only.

A runtime override is possible by placing a configuration into /run,
which masks the distro-provided config. The old location in /etc becomes
the highest priority override and is left to the system administrator.

Putting RAUC along with its configuration into /usr is called hermetic
/usr. The search path order implemented here is the recommendation
described in this patch is aligned with the UAPI group's Configuration
Files Specification[1].

The specification goes beyond what is implemented here and describes
also support for drop-in extensions to the configuration files.

This involves more work, because GKeyFile doesn't support merging of
config files, so drop-in support is skipped for now.

[1]: https://uapi-group.org/specifications/specs/configuration_files_specification/

Closes: https://github.com/rauc/rauc/issues/1252
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>