data/rauc.service.in: set --mount to /run/rauc/mnt subdirectory
In a follow-up commit, we will add /run/rauc to the search path used to
locate the system.conf. To make use of this, the user will have to create
/run/rauc/system.conf prior to RAUC startup and the systemd unit doesn't
mind that.

When the rauc.service restarts though, RuntimeDirectory=rauc in the
service unit will cause systemd to remove the directory and along with
it /run/rauc/system.conf.

Avoid this by using /run/rauc/mnt as mount directory. systemd will only
delete the mnt subdirectory and keep /run/rauc as-is.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>