src/context.c: refactor to reduce code duplication
The current code is a bit complicated because it needs to handle three
different cases:

  - config file is required
  - config file should be used if available, otherwise defaults
  - defaults should be used from the get-go

As a result, it duplicates loading the config, error check, assignment of
context->config and calls default_config() twice at different places.

By splitting out the common parts into a helper function, the code is
arguably easier to follow.

The only intended functional change of this commit is always writing out
where a non-default config came from. We didn't do this before with
configmode == R_CONTEXT_CONFIG_MODE_REQUIRED, but with incoming support
for looking up the RAUC config at different locations, it makes sense to
always report which config RAUC ultimately decided to use.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>