qemu-test-init: don't touch test/grubenv.test
Not a single test is actually using test/grubenv.test as created here.
Thus skip the indirection of creating and bind-mounting it and just care
for the qemu grubenv.
This avoids confusion and wrong expectations.

Also omit the grub-editenv 'create' call which is not necessary since
'set' will create the env file if necessary.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>