src/install: Check slot status to skip up-to-date slots
When an installation of a slot with install-same=false fails, the initial rauc install attempt will fail.
However, a subsequent attempt with the same rauc bundle will succeed as the slot is skipped.

This happens because "failed" slots also have a valid checksum set
And the slot status is not checked to be "ok" before comparing the checksum.

Signed-off-by: Samer Kenawy <samer.kenawy@nl.abb.com>