src/update_handler: ignore mount error before running slot hook
The handler mounts the other-slot filesystem, and it returns in error if
it fails to mount it.

The filesystem may be in bad health, or may be not yet made, and thus,
it causes the handler to fail.

As a consequence, it results in a slot that is unupdatable until the
filesystem is remade manually (i.e. externally).

This ignores the mount error before the slot hook is run, and leaves the
hook to decide whether or not to fail.

Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>