src/bundle: casync_make_arch(): don't use fakeroot in pseudo
When pseudo is active, like it would be in a pseudo-enabled bitbake
task, calling fakeroot is useless or can cause issues.

For this purpose, deactivate the call to 'fakeroot' when detecting an
active pseudo environment.

We can simply leverage the existing check_pseudo_active() method which
was introduced with adaptive update conversion handling.

While simply omitting the 'fakeroot' part of the call might looks a
little hacky, it is the least invasive solution to the problem.

A cleaner, but more invasive solution would be to re-use the fakeroot
env handling based on r_fakeroot_*() methods. However, this only works
with true fakeroot and will fail when being called from a 'pseudo'
acting as a 'fakeroot' like it is done in the 'meta-rauc' OE layer.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>