qemu-test-init: use INTERACTIVE_SHELL instead of SHELL
TL;DR; This fixes an issue on systems which use bash as sh.

The argument shell passed through the kernel argument to qemu-test-init
sets SHELL=1 locally, but...

SHELL is an environment variable containing the path to the shell; it is
set by login(1) and probably many other programs such as sudo(8).

Unlike the other shells (zsh, dash, ash, fish...), bash ensures SHELL if
set[1]!

Since qemu-test-init is a shell script, it starts with SHELL=/bin/bash
set in its environment if bash(1) is used to run the script, i.e. if
bash is the default shell /bin/sh.

This causes the interactive shell to always run even if the shell
argument is not part of the kernel commandline (i.e. if the shell
argument is not given to qemu-test).

Therefore, SHELL is a bad name candidate to locally store in the script
if an interactive shell is to be run or not.

This renames the local variable SHELL to INTERACTIVE_SHELL to prevent
from always running an interactive shell on system using bash as the
default shell.

[1]: https://git.savannah.gnu.org/cgit/bash.git/tree/variables.c?h=bash-5.2#n758

Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>