qemu-test-init: mark as good before starting shell
The argument system sets SHELL and SERVICE to run an interactive shell
after the RAUC service is started.

It was introduced by the commit f495924752c558076baf09d431e473e6bc448041
and the mark good happens once the interactive shell ends.

But, the shell termination shuts down the system and the mark good has
no strong sense to happen before the system powers off, if only it has
the time to happens.

This moves the mark as good before starting the interactive shell.

Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>