qemu-test-init: fix testing the entire test suite
The documentation says qemu-test runs the entire test suite[1] if the
test= argument is not given. But, the log attached below reports an
error.

The argument test=subprojname:testname was introduced by the commit
cfa5260b7c9b62d625884abfd02bed8407712534 to run specific tests only.

If test=bootchooser is given to qemu-test, qemu-test-init runs meson -C
build "bootchooser" -v and it runs the test.

If no test= is given, qemu-test-init runs meson -C build "" -v and if
fails running the whole test suite because the testname is an empty
string:

	ERROR: *: test name does not match any test

This fixes the testing for the whole test suite by setting the script
positional parameters to "$test" and by using "$@" to expand the
specified testname with the meson-test command:
 - if test= is set, "$@" expands to the test= value
 - otherwise, if test= is unset, "$@" expands to nothing

Fixes:

	+ meson test -C /home/gportay/src/rauc/build/ '' -v

	ERROR: *: test name does not match any test
	+ cat /home/gportay/src/rauc/build//meson-logs/testlog.txt
	cat: /home/gportay/src/rauc/build//meson-logs/testlog.txt: No such file or directory
	(...)
	+ echo 'RESULT: FAILED'
	RESULT: FAILED

[1]: https://rauc.readthedocs.io/en/v1.13/contributing.html#qemu-test-runner-qemu-test

Co-authored-by: Jan Lübbe <jluebbe@lasnet.de>
Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>