src/mount: provide more helpful error message when squashfs is missing
According to the mount() syscall manpage, ENODEV means
"filesystemtype not configured in the kernel."

Since we know our file system, we can directly inform the user about the
missing 'squashfs' support in kernel.

While at it, also follow the rule of not adding information a caller
would know anyway and remove the "failed to mount bundle: " prefix.

So instead of:

| Failed mounting bundle: failed to mount bundle: No such device

we get

| Failed mounting bundle: squashfs support not enabled in kernel

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>