src/artifacts: add error handling for tree repo type
E.g. when using the 'composefs' convert method, the original file will
not be preserved.
When accidentally attempting to install this image into a 'tree' type
repo, the g_file_test() check would fail and the 'converted' filename
would be NULL when passed to tree_artifact_install_tar() (and cause an
assertion error).

Better inform the user explicitly that he attempts to install an
artifact with the wrong repo type.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>