add support for composefs artifacts
composefs combines overlayfs, EROFS and optionally fs-verity into an
authenticated read-only filesystem with content-addressed file data. In
RAUC, we can use it to optimize storage and download of artifiacts such
as container filesystems. This commit integrates it as a new repository
type with a corresponding conversion method.

The use-cases for this are very similar to the existing ``tree`` type,
but as composefs stores file data separately from metadata, every unique
file content is stored in the repository only once and is installed from
the bundle only once. This also reduces the amount of data downloaded
when using streaming.

In the future, this type could be extended to reuse file data from the
normal slots and with delta compression for large files. Also, it may be
useful to add support to optionally enable fs-verity for the objects and
metadata images in a composefs repo.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>