add support for directories in bundles
In some cases, it's useful to store additional files in the bundle which
are not used directly by RAUC. To keep things organized, it can be
useful to place them in subdirectories.

As RAUC may need to store it's own data in subdirectories (e.g.
'.rauc-cfs-store' for composefs objects), hidden directories stay
reserved for internal use.

Signed-off-by: Jens Kehne <jens.kehne@agilent.com>
Co-developed-by: Jan Luebbe <jlu@pengutronix.de>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>