src: simplify and fix NFS (external) boot detection
Since 2928ac06 ("src/main: fix and add support for boot detection with
rauc.external"), we consider "_external_" in r_context()->bootslot as an
indicator for an external boot.
While this works well for explicit 'rauc.external' boots, it does not
for boots with 'root=/dev/nfs' in the commandline since
r_context()->bootslot is '/dev/nfs' in this case.

Unify these cases by returning "_external_" from get_cmdline_bootname()
both for the 'rauc.external' as well as the 'root=/dev/nfs' case.

To not loose the info printout for external or nfs, move these to
get_cmdline_bootname(), too.

From here on it is now actually safe to use "_external_" as an
indicator for all kinds of external boots.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>