src/mount: open /dev/loop* readonly
We don't intend to write to the RAUC bundle, so open /dev/loop*
readonly.

Mounting readwrite only works when the Kernel option
CONFIG_BLK_DEV_WRITE_MOUNTED is enabled. Without this option
mounting the bundle fails with -EBUSY.

With CONFIG_BLK_DEV_WRITE_MOUNTED disabled the Kernel will forbid
writing to a block device or file while being mounted to avoid
filesystem corruptions. Without this patch we have /dev/loop* opened
readwrite. Consequently the Kernel forbids mounting it, because that
would result in a situation where we could write to the underlying file
through the file descriptor. This only happens for 'plain' bundles.

Technically we could close the loop device file descriptor before
mounting the loop device because we don't need it anymore. However, we
have set the LO_FLAGS_AUTOCLEAR flag on the loop device which makes it
disappear once we close the file descriptor.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>