src/context: custom-bootloader: error if get-current outputs nothing
The custom bootloader ends with the error below if it fails to find any
root device or rauc slot information from the /proc/cmdline, and if the
optional action get-current outputs an empty string:

	[root@qemu-test rauc]# rauc service
	rauc-Message: 22:17:49.242: Using central status file /tmp/rauc/central.raucs
	rauc-Message: 22:17:49.244: Using system config file /etc/rauc/system.conf
	rauc-Message: 22:17:49.245: Failed to load system status: No such file or directory
	rauc-Message: 22:17:49.262: Failed to resolve realpath for '/dev/root'
	Failed to determine slot states: Did not find booted slot (matching '')
	                                                                    ^^

The error is not explicit as the function r_boot_get_current_bootname()
returns an empty string satisfying r_context_configure_target().

The setup continues and fails in the function determine_slot_states()
after it fails to find a slot that match an empty bootname.

This returns an error in the function custom_backend_get_bootname() if
the optional action get-current outputs nothing, making the error
slightly more explicit. The error happens a few lines earlier in the
same function determine_slot_states():

	[root@qemu-test rauc]# rauc service
	rauc-Message: 22:18:57.760: Using central status file /tmp/rauc/central.raucs
	rauc-Message: 22:18:57.767: Using system config file /etc/rauc/system.conf
	rauc-Message: 22:18:57.769: Failed to load system status: No such file or directory
	rauc-Message: 22:18:57.811: Failed to get bootname: custom backend: Failed to get custom backend bootname '/etc/rauc/custom-bootloader-script': no output
	Failed to determine slot states: Could not find any root device or rauc slot information in /proc/cmdline

Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>