src/manifest: ignore the rollout section for now
The main polling configuration is set in the system.conf. In the future,
we may want to further configure the rollout, e.g. to specific time
windows or for phases.

By ignoring unknown options in this section until then, it will be
easier to use them later as an intermediate update can be avoided.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>