add support for requiring a specific manifest hash during installation
Check that the hash of the to be installed bundle's manifest matches the
expected value. This can be used to when explicit confirmation is needed
before installing a bundle by first reviewing the information returned
by InspectBundle and then passing the hash here. If the bundle was
replaced by a different (but correctly signed) bundle, this is detected
by comparing the manifest hashes.

This is also useful with the upcoming polling support if the user wants
to automate the polling only and still trigger the installation
explicitly.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>