src/context: priorize bootchooser to select bootslot
The commit 9af0613c7527ebadc855c1a49a4d10757150194a introduced the new
bootchooser API r_boot_get_current_bootname() to allow the backend to
determine the current bootslot.

Since then, the bootchooser is able determine the bootslot if it cannot
be guessed from the kernel cmdline; it is used by the custom backend
only at that moment.

This gives the priority to the bootchooser over the kernel cmdline.