implement polling in the service
This implements large parts of #1114 (Simple Update Server Support).

The polling functionality in RAUC allows a device to periodically check
for updates from a specified source. It fetches bundle manifests, checks
for available updates, and initiates installations if necessary. This
functionality is intended to automate the update process and ensure that
the system remains up-to-date with minimal manual intervention.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>