contrib: polling-server: example on bundle HTTP headers and HTTP redirects
RAUC can be configured to send custom HTTP headers using the `send-headers`
parameter in the `[streaming]` section and `RAUC_HTTP_` keys set by
the `system-info` handler script.

These are intended for telemetry collection about devices in the field and
controlled rollouts of devices in the wild by providing different bundles
based on information in the header.

This will become more relevant with the upcoming polling feature,
which enables the automatic installation of updates using RAUC alone.

This script provides a basic example on how to use these features.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>