src/nbd: Set groups of child process
When rauc is streaming a bundle during an installation operation, it
creates a sub-process, and switches the user of this subprocess, to
create e sandbox environment where this user have access only to the
required resources in the system.

Until now, when this subprocess received the UID and GID of the
streaming user, however, the groups of this user were not set
accordingly.
This can cause problems when resources needed for the installation
(e.g., installation keys) are available only through group permissions
of the sandbox user.

This commit adds relevant logic to the child preparation and setup
functions, where the code finds out which the user is a part of and set
them accordingly.

Setting the groups of the process must happen before changing the UID of
the process, otherwise, the group setting operation will fail with
permissions errors.

Signed-off-by: Omri Sarig <omri.sarig@prevas.dk>