src/nbd: create typedef for child_setup_args
Originally, the struct of child_setup_args did not have any typedef, and
was always used as "struct child_setup_args".

In the future, this struct will be surrounded by gmacro macros, to
handle memory management of the fields in this struct. The gmacro macros
are using the type name of the given object when creating their own
objects through these macros.
To make it possible for an object to have these macros, the type of the
object must be constructed of a single word, otherwise, the gmacros will
break by the space in the type name.

The current commit adds the typedef to the struct definition, and
updates all the usage in the code from "struct child_setup_args" to be
just "child_setup_args".

Signed-off-by: Omri Sarig <omri.sarig@prevas.dk>