docs/examples: fix grub.cfg example
Prior to this patch if the A side failed to boot the B side would never
have been tried.

This is because default=0 is used as a sentinel value, while in fact it
is also the case where the A side is selected

I think this happened when someone took the grub.conf from the repo [0]
and tried to remove the rescue menuentry. If the rescue menuentry is
present using default=0 as a sentinel value works, as it will never be
selected unless both sides are broken.

[0]: https://github.com/rauc/rauc/blob/master/contrib/grub.conf

Signed-off-by: Linus <linus.wallgren@netinsight.net>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>
Signed-off-by: Enrico Jörns <ejo@pengutronix.de>