add config option allowed-signer-cns to keyring
This optional parameter allows setting specific
allowed CommonNames which will be verified against
the CommonNames in leaf certificates of a device.

The use case for this is to sign several rauc bundles
with the same root certificate and then distinguish
which device is eligible to use a specific rauc bundle
by setting this new config parameter to a CommonName
of a certificate residing on the device.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>