src/slot: fix mount point detection for symlinks
It may be desirable to specify the device of a slot not as the real
device path, but as a symlink to it (e.g. by using a udev-provided
symlink such as /dev/disk/by-id/... instead of /dev/mmcblk...).

The slot mount point detection failed in these cases because the device
passed to r_slot_find_by_device() is always the real path of the device.

Fix this by checking the realpath() of each slot device in addition to
the specified path.

Signed-off-by: Andreas Mützel <andreas.muetzel@emlix.com>