de.pengutronix.rauc.Installer.xml: use CC0-1.0 license
While we initially intended this file to be freely reusable, as we
considered it not to be copyrightable, it seems that opinions differ and
not having an explicit licenses causes friction.

To clarify and simplify the situation, we use the CC0-1.0 to declare
that it can be used without restrictions.

Fixes: #1713
Reported-by: Bruno Thomsen <bruno.thomsen@gmail.com>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>