config_file: add slots-locking
This optional config option adds the possibility
to lock a slot. It is a global setting and is not
bound to any specific slot.
When activated, the slot that is marked good will
be locked. The slot will be unlocked again, when
the slot is marked active.
The config option does not have any influence on
which slot is bootet, it will just stop decrementing
the remaining attempts counter in the bootloader.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>