config_file: extend prevent-late-fallback with lock
The new option for prevent late fallback adds the possibility
to lock the attempts counter. It is a global setting and is not
bound to any specific slot.
When activated, the counter of the slot that is marked good will
be locked and not decrease and increase anymore.
It will be unlocked again, when the slot is marked active.
The config option does not have any influence on which slot is
bootet, it will just stop decrementing the remaining attempts
counter in the bootloader.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>