barebox: add counter locking to barebox backend
Forward the counter locking option to the barebox bootloader.
So the barebox can then stop decrementing the remaining_attempts
counter.
It inhibits fall-back to a previous version of the system, which
can happen if a system is rebootet too frequently before a slot
is marked good again and the remaining_attempts counter is
incrememented. As a side effect, it inhibits excessive write
cycles on the storage medium.

This also needs changes to barebox, see [1].

[1] https://lists.infradead.org/pipermail/barebox/2025-June/051393.html

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>