src/update_handler: basic progress for block hash index generation in adaptive mode
Depending on the CPU and storage speed, generating the hash index for
the adaptive mode 'block-hash-index' might take a notable amount of
time.

So far, we don't emit any progress information during these steps, which
could give the impression that the installation progress stalls.

With this simple approach, we reserve 10% of the progress for both the
'target_slot' and the 'active_slot' generation.
While this does not emit progress *during* the actual hash index
generation, it adds two benefits:

1. An update between the two hash indexes that might need to be
   generated
2. A progress of already 20% after reading or generating the hash index.
   Which might well reflect the optimization even if the index was
   loaded from a file.

Since 10% is a quite arbitrary value, define it by
R_HASH_INDEX_GEN_PROGRESS_SPAN to allow changing it easily and to
reflect its meaning (also for follow-up code added).

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>