src/context: expclit error when using r_context() with uninitialized context
When r_context() is called without calling r_context_conf() first, the
'context' variable will be NULL and we run into an assertion error.

This might especially happen when setting up new tests that require
the context to be set up.

Be more programmer-friendly here and inform about what to do.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>