src/update_handler: add size_limit slot property
Size allows to limit the slots size for RAUC operations, e.g. if there
is some data at the end of the device that should remain untouched by
RAUC.

Added support to clear_slot() and in check_image_size().

This should be sufficient for now since we clear the device/partition up
to size_limit only and ensuring the image is smaller than 'size_limit'
should also prevent writing more than size_limit bytes during image
copying.

Since we only add it, but not use it, no functional change is intended
so far.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>