signature: implement handling of bundles with multiple signatures
On OpenSSL versions which support CMS_VERIFY_PARTIAL, we can accept a
bundle with multiple signatures even if only one of them can be
verified. We use the config option allow-single-signature=true to enable
this and keep the existing behaviour of requiring successful
verification of all signatures as the default for backwards
compatibility.

This can be used to simplify a key rotation process: Instead of signing
the same bundle contents with the old and new keys separately and
documenting when to use each resulting bundle, a single bundle is enough.

In the future, we can build on the OpenSSL feature to actually require
multiple signatures from different/specific signers.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>