utils: add g_auto support for temporary files
In several places, removing an output file on error is the only reason
why we still use 'goto out'. So, add a helper to handle this via g_auto.

If the file should not be removed simply use g_clear_pointer(&filename,
g_free), so that the cleanup is not triggered.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>