src/update_handler: replace dummy slot by modifying original slot
In img_to_boot_emmc_handler(), instead of creating the dummy
'part_slot', we temporarily set the 'device' of the original slot to the
eMMC boot partition device.

This is a preparation for the follow up refactoring commit.

The main purpose is to allow using the same slot for the hooks as we use
for the actual copying.

To keep the hook env consistent, manually override the RAUC_SLOT_DEVICE
with the original eMMC base device.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>