src/emmc: refactor toggle the active boot partition of an eMMC
Move toggling and checking of active partition to its own function .
The function takes the currently active boot partition from its parameters
and changes the active boot partition to the currently inactive boot partition.
It also validates if the activation was successful, in order to check for a known
ioctl ext_csd bug, see [1].

This encapsulates already existing code into a function and makes it easier to reuse
and avoids duplication in the future.

[1]: https://rauc.readthedocs.io/en/latest/advanced.html#update-bootloader-in-emmc-boot-partitions

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>