src/update_handler: handlers for emmc-boot-linked
We need two handlers for emmc-boot-linked, a default one which handles
the normal everyday update procedure when updating from emmc-boot-linked
slots. Additionally migration for updates to emmc-boot-linked from a
different slot configuration, like boot-emmc, must be handled.
When coming from a different slot configuration it could be that the target
slot for the bootloader update is the currently activated one. This must be
properly handled to allow seamless migration.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>