src/update_handler: handler for emmc-boot-linked
We need a handler for emmc-boot-linked which handles the normal everyday
update procedure when updating from emmc-boot-linked slots.
Migration from different slot configurations, like boot-emmc,
to emmc-boot-linked must be handled.
When coming from a different slot configuration the target slot for
the bootloader update can be the same as the currently activated one.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>