src/context: parse the case cmdline contain root=None
In the case of the cmdline contain root=None, rauc service fail. It is
because the function get_cmdline_bootname_root() does not provide for
this scenario and continue to the end before crashing.
If cmdline contain root=None, the function get_cmdline_bootname_root() should
give the same result as for case root=xxx is not define, namely returning NULL
to avoid rauc failures.
root=None on cmdline can occur on Yocto images for x86_64 system.

Signed-off-by: Benjamin Missey <benjamin.missey@non.se.com>