test/test_syslog: verify optional syslog support
This new test verifies syslog support by creating a dummy listener for
the /dev/log socket and checking the log format sent to this socket to
ensure it contains rauc[PID]: style logs like 'service start'.

Three different variations, for maximum coverage, are run:

 --syslog           (no argument, default facility)
 --syslog=local0    (custom facility, should pass OK)
 --syslog=undefined (invalid facility, should FAIL)

Note, test is skipped if /dev/log already exists or if it cannot bind
the domain socket to /dev/log.  Either a syslog daemon is running, a
case the test is not designed for, or the test is running in a cross
environment without podman being started with --privileged.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>